+++ src/main/java/org/example/spring/security/config/SecurityConfig.java
@@ -8,8 +8,8 @@ import org.example.spring.domain.member.MemberRole;
 import org.example.spring.security.handler.CustomAccessDeniedHandler;
 import org.example.spring.security.handler.CustomAuthenticationEntryPoint;
 import org.example.spring.security.jwt.CookieService;
-import org.example.spring.security.jwt.JwtAuthenticationService;
-import org.example.spring.security.jwt.JwtValidatorFilter;
+import org.example.spring.security.service.JwtAuthenticationService;
+import org.example.spring.security.filter.JwtValidatorFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
