+++ src/main/java/org/example/spring/security/service/JwtAuthenticationService.java
@@ -1,8 +1,10 @@
-package org.example.spring.security.jwt;
+package org.example.spring.security.service;
 
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Arrays;
 import java.util.List;
+import org.example.spring.security.jwt.JwtTokenProvider;
+import org.example.spring.security.jwt.JwtTokenValidator;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
